requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
certifi==2024.12.14      
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0     
ecdsa==0.19.0
exceptiongroup==1.2.2    
executing==2.1.0
fastapi==0.115.6
h11==0.14.0
icecream==2.1.4
idna==3.10
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
python-jose==3.3.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
termcolor==2.5.0
